Refactor agent decorator to require Agent protocol
As it turns out, in mypy you cannot alter the type of a class using a
decorator on the class. That was my intention with the @agent
decorator---take any class and make it implement the Actor protocol.
That doesn't work so now the Agent protocol is simplified so users can
implement it in their classes and the @agent decorator will be used for
other things.

This also adds a simple AgentRunner to validate using an Agent works.