Refactor agents and sketch exchange/launcher
Introduces the concept of a Behavior, a class which defines the behavior
of an Agent. What previously was the AgentRunner is now just the
Agent---initialized with a Behavior. This helps to separate the
behavioral logic of an agent with it's execution details.

This refactor came about by sketching the Mailbox, Identifer, Exchange,
and Launcher interfaces and providing simple multi-threading
implementations (untested!).