Refactor agent handles
The prior Handle implementation had a key problem that every handle,
whether held by a client or a running agent, has its own client ID and
mailbox. This meant an agent with handles to 10 other agents had 11
different mailboxes and listening threads rather than just 1. This
doesn't scale well and also the interface made it challenging the create
"handle cycles" between agents (i.e., agent A has a handle to agent B
and vice versa). The last challenge with prior Handles is that it was
hard to test a Behavior in isolation if it was initialized with a Handle
to a running agent.

This commit completely changes the RemoteHandle to have three variants,
unbound, client-bound, and agent-bound so that running agents can bind
handles to their already existing mailbox.

To test this functionality, I had to fix Issue #9 to execute actions in
threads otherwise an action that invoked a remote action would block
forever.