Fix crash/hang in test_launch_agents_processes
The test_launch_agents_processes in tests/launcher/executor_test.py
was not actually testing anything for two reasons:
  1. The agent process was immediately crashing because the Agent
     class was not pickleable.
  2. The message listener thread in the SimpleExchange class never
     exited because it was not closed by the agent so the agent's
     process never exited cleanly, causing the ProcessPoolExecutor to
     hang on shutdown.

The first first was to implement __reduce__ for Agent.

The second fix required uncovering which thread was hanging, so I added
names to every spawned thread for debugging help. Since it was the
thread in the SimpleExchange, I modified the Agent to close it's
exchange when it finished running. This could be a little dangerous in
tests when multiple agents share the same exchange object.